added indicator
the first algorithm cell contains the version that seems to do nothing and then all of a sudden gets a s=NaN as step length. After that there is a new draw function (can you maybe upload yours?) but I don't want to try fixing the "range step cannot be zero" error (that comas from Makie). Last cell shows how indicator and using that works. I'll try implementing this in the main program